requirements.txt
mysql.connector
fastapi
